Rename actor.py to backend.py to be more in line with other plugins
